requirements.txt
Appium-Python-Client
pytest-ordering
selenium